+++ packages/EventMap/src/applyEventMapLogging.ts
@@ -1,19 +1,39 @@
 import {EventMap, EventMapHandlerInfo} from "./EventMap";
-import {Logger} from "@harbor/middleware";
+import {Logger} from "@domx/middleware";
 export { applyEventMapLogging };
 
+let isApplied = false;
 
-const applyEventMapLogging = () => EventMap.applyMiddleware(
-    (handlerInfo:EventMapHandlerInfo) =>
-    (next:Function) => () => {        
-    const el = handlerInfo.element;
-    const detail:any = handlerInfo.eventDetail || "(none)";
 
-    Logger.log(el,"group",
-        `> EVENTMAP: ${handlerInfo.listenAt}@${handlerInfo.eventName} => ` +
-        `${handlerInfo.constructorName}.${handlerInfo.eventHandlerName}()`);
+const applyEventMapLogging = (
+    {collapsed}:{collapsed:boolean} = {collapsed:false}) => {
+        if (isApplied) {
+            console.warn("applyEventMapLogging has already been called.");
+            return;
+        }
+        isApplied = true;
+        
+        EventMap.applyMiddleware(
+            (handlerInfo:EventMapHandlerInfo) =>
+            (next:Function) => () => {     
 
-    Logger.log(el, "info", `=> event.detail`, detail);
-    next();
-    Logger.log(el, "groupEnd");
-});
+            const el = handlerInfo.element;
+            const detail:any = handlerInfo.eventDetail || "(none)";
+            debugger;
+            Logger.log(el, collapsed ? "groupCollapsed": "group",
+                `> EVENTMAP: ${handlerInfo.listenAt}@${handlerInfo.eventName} => ` +
+                `${handlerInfo.constructorName}.${handlerInfo.eventHandlerName}()`);
+
+            Logger.log(el, "info", `=> event.detail`, detail);
+            next();
+            Logger.log(el, "groupEnd");
+        }
+    )
+};
+
+/**
+ * Exposed for testing
+ */
+applyEventMapLogging.reset = () => {
+    isApplied = false;
+};