+++ packages/StateChange/src/test/testHelpers.ts
@@ -11,7 +11,8 @@ interface FixtureElement extends HTMLElement {
     testSimple: Function,
     testError: Function,
     testTapError: Function,
-    testDispatchEvent: Function
+    testDispatchEvent: Function,
+    testImmer: Function
 };
   
   function fixture(html:TemplateResult): FixtureElement {
@@ -37,7 +38,7 @@ class TestStateChange extends HTMLElement {
     testSimple() {
       StateChange.of(this)
         .next(setFooTrue)
-        .next(setBarTrue)
+        .next(setBar(true))
         .dispatch();
     }
   
@@ -81,45 +82,65 @@ class TestStateChange extends HTMLElement {
           composed: true
         }));
     }
+
+    testImmer() {
+      StateChange.of(this)
+        .next(changeStateWithMutations)
+    }
 }
 customElements.define("test-state-change", TestStateChange);
   
   
   
 const setFooTrue = (state:any) => {
-    return {
-      ...state,
-      foo: true
-    };
-  }
+  return {
+    ...state,
+    foo: true
+  };
+};
   
-  const setBarTrue = (state:any) => {
-    return {
-      ...state,
-      bar: true
+const setBarTrue = (state:any) => {
+  return {
+    ...state,
+    bar: true
+  };
+};
+
+
+// for testing inner function name
+const setBar = (bar:boolean) =>
+    function setBar(state:any) {
+      return {
+        ...state,
+        bar
+      };
     };
-  }
   
-  const asyncTest = async (stateChange:StateChange) => {
-    await setTimeout(()=> {}, 10);
+const asyncTest = async (stateChange:StateChange) => {
+    await setTimeout(()=> {});
     stateChange
       .next(setFooTrue)
       .next(setBarTrue)
       .dispatch();
   };
   
-  const setFooTrueWithTap = (stateChange:StateChange) => 
+const setFooTrueWithTap = (stateChange:StateChange) => 
     stateChange.next(setFooTrue);
   
-  const setBarTrueWithTap = (stateChange:StateChange) => 
+const setBarTrueWithTap = (stateChange:StateChange) => 
     stateChange.next(setBarTrue);
   
-  const causeError = (state:any) => {
+const causeError = (state:any) => {
     throw new Error("test error")
   };
 
 
-  const tapError = (stateChange:StateChange) =>
+const tapError = (stateChange:StateChange) =>
     stateChange.next(causeError);
   
+
+const changeStateWithMutations = (state:any) => {
+  state.bar = true;
+  state.newArray = [1];
+};